Updating make script.

+++ make_figs.sh
@@ -12,7 +12,7 @@ export ASYMPTOTE_DIR="$(pwd)/"
 rm -f *.pre *.prc *.js *.tex *.pdf *.png
 
 # Create gallery pages for the images.
-for filename in ./*.asy; do
+for filename in `find . -name "*.asy"`; do
 
     # Extract the file name (returns file from file.asy).
     file=$(basename -- "$filename")
@@ -29,21 +29,18 @@ for filename in ./*.asy; do
         # Remove excess junk.
         rm -f *.pre *.prc *.js *.tex *.pdf *.svg *.eps *.aux *.dvi *.log
 
-        asy "$file.asy"
-        mv "$file.pdf" "$Destination"
+        asy "$filename"
+
+        if [ -e "$file.pdf" ]; then
+            mv "$file.pdf" "$Destination"
+        else
+            mv "$file.eps" "$Destination"
+        fi
     fi
 done
 
 # Remove excess junk.
 rm -f *.pre *.prc *.js *.tex *.pdf *.svg *.eps *.aux *.dvi *.log
-
-CompilerArgs1="-std=c89 -ansi -pedantic -pedantic-errors -Wall -Wextra"
-CompilerArgs2="-Wpedantic -Wmisleading-indentation -Wmissing-field-initializers"
-CompilerArgs3="-Wmissing-prototypes -Wold-style-definition -Winit-self"
-CompilerArgs4="-Wmissing-declarations -Wnull-dereference -Wwrite-strings"
-CompilerArgs5="-Wconversion -Wdouble-promotion -Wstrict-prototypes -O3"
-CompilerArgs="$CompilerArgs1 $CompilerArgs2 $CompilerArgs3"
-CompilerArgs="$CompilerArgs $CompilerArgs4 $CompilerArgs5"
 
 # Execute the C programs.
 cd "$top/programs/"
@@ -63,16 +60,12 @@ for filename in ./*.c; do
     # Compile and execute.
     gcc $CompilerArgs $file.c -o $file -lm
     ./$file
-    if [ "${file: -6}" == "Raster" ] ; then
-        convert "$file.ppm" "$file.png"
-        mv "$file.png" "$Destination"
-    elif [ "${file: -6}" == "Vector" ] ; then
-        convert "$file.svg" "$file.pdf"
-        mv "$file.pdf" "$Destination"
+    if [ -e "$file.pdf" ]; then
+        mv $file.pdf $Destination
+    elif [ -e "$file.svg" ]
+        mv $file.svg $Destination
     else
-	    echo "Last characters should be 'Raster' or 'Vector'"
-	    echo "Aborting."
-	    exit 1
+        mv $file.ppm $Destination
     fi
-    rm -f "$file" "$file.ppm" "$file.svg"
+    rm -f $file
 done
